+++ app/services/express/Strategy/facebook.js
@@ -1,0 +1,26 @@
+import FacebookStrategy from 'passport-facebook';
+import Users from '../../../api/users/model';
+let secrets = require('./secrets.json');
+
+function fbcallback(token, refreshToken, profile, done) {
+    // asynchronous
+    process.nextTick(function () {
+        // find the user in the database based on their facebook id
+        Users.findOne({'id': profile.id})
+            .then(function (user) {
+                // if the user is found, then log them in
+                if (user) {
+                    done(null, user); // user found, return that user
+                } else {
+                    // save our user to the database
+                    Users.insert(profile._json).then((newUs) => done(null, newUs));
+                }
+
+            })
+            .catch(done);
+    });
+}
+
+export default new FacebookStrategy(secrets.facebook,
+    // facebook will send back the token and profile
+    fbcallback);
