+++ app/services/express/Strategy/local.js
@@ -1,23 +1,15 @@
 import LocalStrategy from 'passport-local';
-import bcrypt from 'bcrypt';
-import Users from '../../api/users/model';
+import Users from '../../../api/users/model';
+import {validatePassword, generateHash} from '../../node/pass-hash';
 
-function generateHash(password) {
-    return bcrypt.hash(password, bcrypt.genSaltSync(10));
-}
-
-function validPassword(password, hash) {
-    return bcrypt.compare(password, hash);
-}
-
-const localStrategyParameters = {
+const params = {
     usernameField: 'email',
     passwordField: 'password',
     passReqToCallback: true
 };
 
 
-let localStrategy = (req, email, password, done) => {
+let strategy = (req, email, password, done) => {
     Users.findOne({email})
         .then((user) => {
             if (!user) {
@@ -34,9 +26,9 @@ let localStrategy = (req, email, password, done) => {
                     });
 
             } else {
-                validPassword(password, user.hashPassword)
-                    .then((isValid) => {
-                        if (!isValid) {
+                validatePassword(password, user.hashPassword)
+                    .then((valid) => {
+                        if (!valid) {
                             done(null, false);
                         } else {
                             done(null, user);
@@ -47,4 +39,4 @@ let localStrategy = (req, email, password, done) => {
         .catch(done);
 };
 
-export default new LocalStrategy(localStrategyParameters, localStrategy);
+export default new LocalStrategy(params, strategy);
