+++ app/services/express/Strategy/serialze.js
@@ -1,4 +1,4 @@
-import Users from '../../api/users/model';
+import Users from '../../../api/users/model';
 
 let serialize = (user, done) => done(null, user.id);
 let deserialize = (id, done) => Users.findOne({id}, done);
